Open Talk url when feedback closes Update the Done & Talk onClick handler to set the url redirect as the onHide callback for feedback.
